make command weights double and tweak some weights

+++ src/main/kotlin/juandb/entelect/challenge/command/Command.kt
@@ -3,6 +3,6 @@ package juandb.entelect.challenge.command
 import juandb.entelect.challenge.entity.GameState
 
 abstract class Command(protected val gameState: GameState) {
-	abstract fun getWeight(): Int
+	abstract fun getWeight(): Double
 	abstract fun getCommand(): String
 }
